+++ frontend/src/components/mod.rs
@@ -8,6 +8,7 @@ mod question_popup;
 mod share_popup;
 mod socket;
 mod spinner;
+mod textarea;
 mod upgrade;
 
 pub use delete_popup::DeletePopup;
@@ -19,4 +20,5 @@ pub use question_popup::QuestionPopup;
 pub use share_popup::SharePopup;
 pub use socket::{EventSocket, SocketResponse};
 pub use spinner::Spinner;
+pub use textarea::TextArea;
 pub use upgrade::Upgrade;